# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API PedidosYa",
        "description": "Esta API puede usarse para obtener información sobre los items y tipos, también consultar pedidos propios y crear un pedido.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/rest/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "description": "Obtiene los detalles de un item específico por su ID.",
                "operationId": "52671cf08305d5d9d4254560adfb3559",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DetalleItem"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item no encontrado"
                    }
                }
            }
        },
...
        "/rest/tipos/{id}": {
            "get": {
                "tags": [
                    "Tipos"
                ],
                "description": "Obtiene una lista de items por tipo según su ID.",
                "operationId": "getItemsPorTipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de items por tipo obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Item"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Item": {
                "properties": {
                    "current page": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "nombre": {
                                    "type": "string",
                                    "example": "Gomitas Mogul"
                                },
                                "etiqueta": {
                                    "type": "string",
                                    "example": "Snack"
                                },
                                "precio": {
                                    "type": "integer",
                                    "example": 140
                                },
                                "activo": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "path_imagen": {
                                    "type": "string",
                                    "example": "images/Gomitas_Mogul.png"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "first_page_url": {
                        "type": "string",
                        "example": "http://digital-dynamos-laravel.vercel.app/rest/items?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page_url": {
                        "type": "string",
                        "example": "http://digital-dynamos-laravel.vercel.app/rest/items?page=1"
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "type": "stringr",
                                    "example": "http://digital-dynamos-laravel.vercel.app/rest/items?page=1"
                                },
                                "label": {
                                    "type": "integer",
                                    "example": "1"
                                },
                                "active": {
                                    "type": "boolean",
                                    "example": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": null
                    },
                    "path": {
                        "type": "string",
                        "example": "http://digital-dynamos-laravel.vercel.app/rest/items"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": null
                    },
                    "to": {
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "DetalleItem": {
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Coca-Cola"
                    },
                    "etiqueta": {
                        "type": "string",
                        "example": "bebida no alcoholica"
                    },
                    "tamaño": {
                        "type": "string",
                        "example": "mediano"
                    },
                    "coste": {
                        "type": "integer",
                        "example": 600
                    },
                    "path_imagen": {
                        "type": "boolean",
                        "example": "images/Coca-Cola.png"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}